Extract OrderDetailsContext from CheckoutContext
