+++ src/Sylius/Behat/Resources/config/services/contexts/ui.xml
@@ -276,7 +276,6 @@
             <argument type="service" id="sylius.behat.page.shop.checkout.select_payment" />
             <argument type="service" id="sylius.behat.page.shop.checkout.select_shipping" />
             <argument type="service" id="sylius.behat.page.shop.checkout.complete" />
-            <argument type="service" id="sylius.behat.page.shop.order.show" />
             <argument type="service" id="__symfony__.sylius.factory.address" />
             <argument type="service" id="sylius.behat.current_page_resolver" />
             <argument type="service" id="__symfony__.sylius.address_comparator" />
@@ -285,6 +284,11 @@
 
         <service id="sylius.behat.context.ui.checkout.thank_you" class="Sylius\Behat\Context\Ui\Shop\Checkout\ThankYouContext">
             <argument type="service" id="sylius.behat.page.shop.order.thank_you" />
+            <tag name="fob.context_service" />
+        </service>
+
+        <service id="sylius.behat.context.ui.checkout.order_details" class="Sylius\Behat\Context\Ui\Shop\Checkout\OrderDetailsContext">
+            <argument type="service" id="sylius.behat.page.shop.order.show" />
             <tag name="fob.context_service" />
         </service>
